requirements.txt
asn1crypto
astroid
beautifulsoup4
certifi
cffi
chardet
cloudscraper
cryptography
python-dotenv
future
idna
isort
lazy-object-proxy
lxml
mccabe
pycparser
pylint
pyparsing
python-telegram-bot
PyYAML
requests
requests-toolbelt
six
soupsieve
typed-ast
urllib3
wrapt
selenium
sqlalchemy==2.0.21
python-dotenv==1.0.0
psycopg2-binary==2.9.7
webdriver_manager
termcolor